chore: refactor static content
