+++ src/app/page.tsx
@@ -10,8 +10,8 @@ import Header from '@/components/header';
 import HomeAvailableTools from '@/components/home/home-available-tools';
 import HomeProjectCarousel from '@/components/home/home-project-carousel';
 import HomeTools from '@/components/home/home-tools';
+import { projects } from '@/data/projects';
 import useScreenSize from '@/hooks/useScreenSize';
-import { projects } from '@/services/project.service';
 
 import 'aos/dist/aos.css';
 import '@/styles/wave.css';
