+++ src/components/auth/forgot-password-modal-content/index.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 
-import authHeaderContent from '@/libs/auth-header-content';
+import { authHeaderContent } from '@/data/authHeaderContent';
 import AuthType from '@/types/auth-type-enum';
 
 import AuthModalHeading from '../auth-modal-heading';
