+++ src/components/auth/sign-up-modal-content/index.tsx
@@ -1,6 +1,6 @@
 import { useSearchParams } from 'next/navigation';
 
-import authHeaderContent from '@/libs/auth-header-content';
+import { authHeaderContent } from '@/data/authHeaderContent';
 import AuthType from '@/types/auth-type-enum';
 
 import AuthModalHeading from '../auth-modal-heading';
