+++ src/components/nav/index.tsx
@@ -5,8 +5,8 @@ import { useContext } from 'react';
 
 import CloseButton from '@/components/close-button';
 import AuthContext from '@/contexts/authContext';
+import { navItems } from '@/data/navItems';
 import useScreenSize from '@/hooks/useScreenSize';
-import { navItems } from '@/libs/api/nav-items';
 import ScreenSize from '@/types/screen-size-enum';
 
 import Button from '../button';
