+++ src/components/sidebar/index.tsx
@@ -6,8 +6,8 @@ import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import { useState } from 'react';
 
+import { sidebarItems } from '@/data/sidebarItems';
 import useScreenSize from '@/hooks/useScreenSize';
-import { sidebarItems } from '@/libs/sidebar-content';
 import ScreenSize from '@/types/screen-size-enum';
 
 type SidebarItemProps = {
