+++ src/types/project-type.ts
@@ -1,0 +1,7 @@
+export type Project = {
+  name: string;
+  description: string;
+  imgSrc: string;
+  imgAlt: string;
+  url: string;
+};
