+++ src/types/tool-type.ts
@@ -1,0 +1,11 @@
+export type Tool = {
+  name: string;
+  description: string;
+  iconFilenames: string[];
+  secondaryIconClass: string;
+  primaryIconClass?: string;
+  decorImgSrc: string;
+  imgAlt: string;
+  active: boolean;
+  url: string;
+};
